// src/app/models/statistique.ts
export class Statistique {
    m1: string;
    m2: string;
    m3: string;
    m4: string;
    m5: string;
    m6: string;
    m7: string;
    m8: string;
    m9: string;
    m10: string;
    
    nb1M1: number;
    nbNM1: number;
    nb2M1: number;

    nb1M2: number;
    nbNM2: number;
    nb2M2: number;

    nb1M3: number;
    nbNM3: number;
    nb2M3: number;

    nb1M4: number;
    nbNM4: number;
    nb2M4: number;

    nb1M5: number;
    nbNM5: number;
    nb2M5: number;

    nb1M6: number;
    nbNM6: number;
    nb2M6: number;

    nb1M7: number;
    nbNM7: number;
    nb2M7: number;

    nb1M8: number;
    nbNM8: number;
    nb2M8: number;

    nb1M9: number;
    nbNM9: number;
    nb2M9: number;

    nb1M10: number;
    nbNM10: number;
    nb2M10: number;

  }